Command Refactor, Cleanup
