+++ velocity/src/main/java/me/leoko/advancedban/velocity/listener/CommandReceiverVelocity.java
@@ -1,4 +1,4 @@
-package me.leoko.advancedban.velocity.command;
+package me.leoko.advancedban.velocity.listener;
 
 import com.velocitypowered.api.command.CommandSource;
 import com.velocitypowered.api.command.SimpleCommand;
